added redis tests
